Organize ConfiguredButton and PearlImage components
