+++ ui-participant/src/landing/sections/HeroCenteredTemplate.tsx
@@ -4,12 +4,12 @@ import React from 'react'
 import ReactMarkdown from 'react-markdown'
 
 import { SectionConfig } from 'api/api'
-import PearlImage, { PearlImageConfig, validatePearlImageConfig } from 'util/PearlImage'
 import { getSectionStyle } from 'util/styleUtils'
 import { withValidatedSectionConfig } from 'util/withValidatedSectionConfig'
 import { requireOptionalArray, requireOptionalString } from 'util/validationUtils'
 
-import ConfiguredButton, { ButtonConfig, validateButtonConfig } from './ConfiguredButton'
+import ConfiguredButton, { ButtonConfig, validateButtonConfig } from '../ConfiguredButton'
+import PearlImage, { PearlImageConfig, validatePearlImageConfig } from '../PearlImage'
 
 type HeroCenteredTemplateConfig = {
   blurb?: string, //  text below the title
