+++ ui-participant/src/landing/sections/PhotoBlurbGrid.tsx
@@ -2,10 +2,11 @@ import React from 'react'
 import ReactMarkdown from 'react-markdown'
 
 import { SectionConfig } from 'api/api'
-import PearlImage, { PearlImageConfig, validatePearlImageConfig } from 'util/PearlImage'
 import { getSectionStyle } from 'util/styleUtils'
 import { withValidatedSectionConfig } from 'util/withValidatedSectionConfig'
 import { requireOptionalArray, requireOptionalString, requirePlainObject, requireString } from 'util/validationUtils'
+
+import PearlImage, { PearlImageConfig, validatePearlImageConfig } from '../PearlImage'
 
 type PhotoBlurbGridConfig = {
   title?: string,
